Fix configuration for development and test
With the current code, on the next step the user won't be able to execute the tests on H2, because the tutorial suggest to change the global database driver.
I propose to change only the 'development' configuration to mysql. This allows the user to use mysql on development and H2 en test.